+++ src/api/groups/projects/projects.rs
@@ -51,7 +51,7 @@ impl GroupProjectsOrderBy {
 }
 
 impl ParamValue<'static> for GroupProjectsOrderBy {
-    fn as_value(self) -> Cow<'static, str> {
+    fn as_value(&self) -> Cow<'static, str> {
         self.as_str().into()
     }
 }