+++ src/api/params.rs
@@ -20,12 +20,12 @@ use crate::api::BodyError;
 pub trait ParamValue<'a> {
     #[allow(clippy::wrong_self_convention)]
     /// The parameter value as a string.
-    fn as_value(self) -> Cow<'a, str>;
+    fn as_value(&self) -> Cow<'a, str>;
 }
 
 impl ParamValue<'static> for bool {
-    fn as_value(self) -> Cow<'static, str> {
-        if self {
+    fn as_value(&self) -> Cow<'static, str> {
+        if *self {
             "true".into()
         } else {
             "false".into()
@@ -34,57 +34,56 @@ impl ParamValue<'static> for bool {
 }
 
 impl<'a> ParamValue<'a> for &'a str {
-    fn as_value(self) -> Cow<'a, str> {
-        self.into()
+    fn as_value(&self) -> Cow<'a, str> {
+        (*self).into()
     }
 }
 
 impl ParamValue<'static> for String {
-    fn as_value(self) -> Cow<'static, str> {
-        self.into()
+    fn as_value(&self) -> Cow<'static, str> {
+        self.clone().into()
     }
 }
 
 impl<'a> ParamValue<'a> for &'a String {
-    fn as_value(self) -> Cow<'a, str> {
-        self.into()
+    fn as_value(&self) -> Cow<'a, str> {
+        (*self).into()
     }
 }
 
 impl<'a> ParamValue<'a> for Cow<'a, str> {
-    fn as_value(self) -> Cow<'a, str> {
-        self
+    fn as_value(&self) -> Cow<'a, str> {
+        self.clone()
     }
 }
 
 impl<'a, 'b: 'a> ParamValue<'a> for &'b Cow<'a, str> {
-    fn as_value(self) -> Cow<'a, str> {
-        let as_ref: &'a str = self.as_ref();
-        as_ref.into()
+    fn as_value(&self) -> Cow<'a, str> {
+        (*self).clone()
     }
 }
 
 impl ParamValue<'static> for u64 {
-    fn as_value(self) -> Cow<'static, str> {
+    fn as_value(&self) -> Cow<'static, str> {
         format!("{}", self).into()
     }
 }
 
 impl ParamValue<'static> for f64 {
-    fn as_value(self) -> Cow<'static, str> {
+    fn as_value(&self) -> Cow<'static, str> {
         format!("{}", self).into()
     }
 }
 
 impl ParamValue<'static> for DateTime<Utc> {
-    fn as_value(self) -> Cow<'static, str> {
+    fn as_value(&self) -> Cow<'static, str> {
         self.to_rfc3339_opts(chrono::SecondsFormat::Secs, true)
             .into()
     }
 }
 
 impl ParamValue<'static> for NaiveDate {
-    fn as_value(self) -> Cow<'static, str> {
+    fn as_value(&self) -> Cow<'static, str> {
         format!("{}", self.format("%Y-%m-%d")).into()
     }
 }