+++ src/api/projects/merge_requests/edit.rs
@@ -22,7 +22,7 @@ enum MergeRequestLabels<'a> {
 }
 
 impl<'a, 'b: 'a> ParamValue<'a> for &'b MergeRequestLabels<'a> {
-    fn as_value(self) -> Cow<'a, str> {
+    fn as_value(&self) -> Cow<'a, str> {
         match self {
             MergeRequestLabels::Unlabeled => "".into(),
             MergeRequestLabels::Labeled(labels) => format!("{}", labels.iter().format(",")).into(),
@@ -49,7 +49,7 @@ impl MergeRequestStateEvent {
 }
 
 impl ParamValue<'static> for MergeRequestStateEvent {
-    fn as_value(self) -> Cow<'static, str> {
+    fn as_value(&self) -> Cow<'static, str> {
         self.as_str().into()
     }
 }